+++ src/v/security/logger.cc
@@ -1,0 +1,17 @@
+/*
+ * Copyright 2024 Redpanda Data, Inc.
+ *
+ * Use of this software is governed by the Business Source License
+ * included in the file licenses/BSL.md
+ *
+ * As of the Change Date specified in that file, in accordance with
+ * the Business Source License, use of this software will be governed
+ * by the Apache License, Version 2.0
+ */
+#include <seastar/util/log.hh>
+
+namespace security {
+
+seastar::logger seclog("security");
+
+}
